pyproject.toml
[project]
name = "vlm-tg-context"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyprojroot>=0.3.0",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "vllm==0.15.1",
    "transformers==4.57.6",
    "ipykernel>=7.1.0",
    "jupyter-server>=2.17.0",
    "jupyterlab>=4.5.4",
    "vulture>=2.15",
    "streamlit>=1.45.0",
    "pytest>=9.0.2",
]
